endeavoru: Switch to /data/media storage layout
Make the switch to the better supported /data/media storage layout.
Due to the small size of our /data partition users can run out of space
pretty quick when using Lollipop (or Kitkat with ART). Moving Apps the
SD card is no long-term solution and the best way seems to be moving to
the now standard /data/media storage layout.

This requires a new recovery and a complete reset (including SD card) of
the phone.

Change-Id: Ida34281cc0e01f21e18300a3a72a7088822ba97d

Conflicts:
	releasetools.py